refactor(v3): split models by domain
