+++ src/maicoin/v3/client.py
@@ -36,38 +36,38 @@ from maicoin.v3.auth import generate_nonce
 from maicoin.v3.errors import Response
 from maicoin.v3.errors import raise_for_api_error
 from maicoin.v3.errors import raise_for_response_status
-from maicoin.v3.models import Account
-from maicoin.v3.models import ConvertOrder
-from maicoin.v3.models import Currency
-from maicoin.v3.models import Deposit
-from maicoin.v3.models import DepositAddress
-from maicoin.v3.models import Depth
-from maicoin.v3.models import FundTransactionDeposit
-from maicoin.v3.models import FundTransactionTransfer
-from maicoin.v3.models import FundTransactionWithdrawal
-from maicoin.v3.models import HistoricalIndexPrice
-from maicoin.v3.models import InterestRate
-from maicoin.v3.models import InternalTransfer
-from maicoin.v3.models import KLine
-from maicoin.v3.models import Market
-from maicoin.v3.models import MWalletADRatio
-from maicoin.v3.models import MWalletInterest
-from maicoin.v3.models import MWalletLiquidation
-from maicoin.v3.models import MWalletLiquidationDetail
-from maicoin.v3.models import MWalletLoan
-from maicoin.v3.models import MWalletRepayment
-from maicoin.v3.models import MWalletTransfer
-from maicoin.v3.models import Order
-from maicoin.v3.models import OrderSide
-from maicoin.v3.models import OrderType
-from maicoin.v3.models import PrivateTrade
-from maicoin.v3.models import PublicTrade
-from maicoin.v3.models import Reward
-from maicoin.v3.models import Ticker
-from maicoin.v3.models import Timestamp
-from maicoin.v3.models import UserInfo
-from maicoin.v3.models import WithdrawAddress
-from maicoin.v3.models import Withdrawal
+from maicoin.v3.models.convert import ConvertOrder
+from maicoin.v3.models.funds import Deposit
+from maicoin.v3.models.funds import DepositAddress
+from maicoin.v3.models.funds import FundTransactionDeposit
+from maicoin.v3.models.funds import FundTransactionTransfer
+from maicoin.v3.models.funds import FundTransactionWithdrawal
+from maicoin.v3.models.funds import InternalTransfer
+from maicoin.v3.models.funds import Reward
+from maicoin.v3.models.funds import WithdrawAddress
+from maicoin.v3.models.funds import Withdrawal
+from maicoin.v3.models.m_wallet import HistoricalIndexPrice
+from maicoin.v3.models.m_wallet import InterestRate
+from maicoin.v3.models.m_wallet import MWalletADRatio
+from maicoin.v3.models.m_wallet import MWalletInterest
+from maicoin.v3.models.m_wallet import MWalletLiquidation
+from maicoin.v3.models.m_wallet import MWalletLiquidationDetail
+from maicoin.v3.models.m_wallet import MWalletLoan
+from maicoin.v3.models.m_wallet import MWalletRepayment
+from maicoin.v3.models.m_wallet import MWalletTransfer
+from maicoin.v3.models.orders import Account
+from maicoin.v3.models.orders import Order
+from maicoin.v3.models.orders import OrderSide
+from maicoin.v3.models.orders import OrderType
+from maicoin.v3.models.orders import PrivateTrade
+from maicoin.v3.models.orders import UserInfo
+from maicoin.v3.models.public_market_data import Currency
+from maicoin.v3.models.public_market_data import Depth
+from maicoin.v3.models.public_market_data import KLine
+from maicoin.v3.models.public_market_data import Market
+from maicoin.v3.models.public_market_data import PublicTrade
+from maicoin.v3.models.public_market_data import Ticker
+from maicoin.v3.models.public_market_data import Timestamp
 
 BASE_URL = "https://max-api.maicoin.com"
 """Default MAX REST v3 base URL."""
