+++ src/maicoin/v3/models/base.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from pydantic import BaseModel
+from pydantic import ConfigDict
+
+
+class MaxBaseModel(BaseModel):
+    model_config = ConfigDict(extra="allow", populate_by_name=True)
